+++ tests/test_functions.py
@@ -2625,9 +2625,6 @@ class TestFunctions(unittest.TestCase):
                          pf.Function('consumption utility',1.,net),
                          pf.Function('net consumption cost',1.,net)]
 
-            # Update network
-            list(map(lambda f: f.update(),functions))
-
             # After updating network
             list(map(lambda f: f.analyze(),functions))
             list(map(lambda f: f.eval(x0),functions))