+++ CHANGELOG.rst
@@ -20,7 +20,9 @@ Unreleased
 * HVDC CSC converters.
 * FACTS devices.
 * Updated count/analyze/eval to also loop through dc buses, and updated python-based dummy func/constr.    
-
+* Changed bus sens_v_reg_by_gen to sens_v_set_reg.
+* Changed "voltage regulation by generators" constraints to "voltage set point regulation".
+  
 Version 1.3.3
 -------------
 * Made Python wrapper its own repo.